surf/pyproject.toml
[project]
name = "surf"
version = "0.1.0"
description = "Self-contained Unified Red-teaming Framework"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "openai>=1.0.0",
    "datasets>=3.5.0",
    "click>=8.0.0",
    "jinja2>=3.1.0",
    "aiofiles>=23.0.0",
    "tqdm>=4.65.0",
    "tenacity>=8.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "torch",
    # Clustering dependencies
    "sentence-transformers>=2.2.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
vllm = ["vllm>=0.6.0"]

[project.scripts]
surf = "surf.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["surf"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
